db exposes en-vi and wordnet
