+++ api/src/db.js
@@ -1,7 +1,7 @@
 const path = require('path')
 const Database = require('better-sqlite3')
 
-exports.getDb = function getDb(dbName, options) {
+function getDb(dbName, options) {
   if (!dbName) throw new Error('Database name is required')
 
   const dbFolder = process.env.NODE_ENV !== 'production' ? 'db-test' : 'db'
@@ -9,3 +9,9 @@ exports.getDb = function getDb(dbName, options) {
   const filename = path.join(process.cwd(), dbFolder, dbName)
   return new Database(filename, { readonly: true, ...options })
 }
+
+const getEnViDb = (options) => getDb('en-vi.db', options)
+
+const getWordnetDb = (options) => getDb('wordnet.db', options)
+
+module.exports = { getDb, getEnViDb, getWordnetDb }
